+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.2] - 2022-07-01
+
+### Added
+
+- Define the sense of the Objective of the StochasticBlock.
+
+### Changed
+
+- Default argument for "issueAMod" parameter in set_data() becomes eNoBlck.
+- Update load() and print() interfaces.
+
 ## [0.4.1] - 2021-12-08
 
 ### Added
@@ -41,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First test release.
 
-[Unreleased]: https://gitlab.com/smspp/stochasticblock/-/compare/0.4.1...develop
+[Unreleased]: https://gitlab.com/smspp/stochasticblock/-/compare/0.4.2...develop
+[0.4.2]: https://gitlab.com/smspp/stochasticblock/-/compare/0.4.1...0.4.2
 [0.4.1]: https://gitlab.com/smspp/stochasticblock/-/compare/0.4.0...0.4.1
 [0.4.0]: https://gitlab.com/smspp/stochasticblock/-/compare/0.3.0...0.4.0
 [0.3.0]: https://gitlab.com/smspp/stochasticblock/-/compare/0.2.0...0.3.0